requirements.txt
algoliasearch==1.9.0
algoliasearch-django==1.2.4
asgi-redis==0.13.1
asgiref==0.13.3
autobahn==0.14.1
cffi==1.7.0
channels==0.14.1
cryptography==1.4
daphne==0.12.1
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.9.7
django-allauth==0.25.2
django-filter==0.13.0
django-redis-cache==1.6.5
djangorestframework==3.3.3
elasticsearch==2.3.0
gunicorn==19.6.0
idna==2.1
msgpack-python==0.4.7
ndg-httpsclient==0.4.1
oauthlib==1.1.1
paypalrestsdk==1.11.7
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.0.0
python3-openid==3.0.10
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
Twisted==16.2.0
txaio==2.5.1
urllib3==1.16
whitenoise==3.2
zope.interface==4.2.0
